lab-02-memory/requirements.txt
langchain>=1.2.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0

